# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Returns scores for the Euro 2024 sweepstakes",
        "title": "Euros Sweepstakes API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/scores": {
            "get": {
                "description": "Get all scores",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scores"
                ],
                "operationId": "get-scores",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/score.Score"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "score.Score": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/score.Team"
                    }
                },
                "totalScore": {
                    "type": "integer"
                }
            }
        },
        "score.Team": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "flag": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        }
    }
}